For direct deposit, don't count as pending edit; go straight to wallet
